proxy/handler: reorganize program startup a little bit
