--- src/proxy/main.h
@@ -1,8 +1,0 @@
-#ifndef PROXY_MAIN_H
-#define PROXY_MAIN_H
-
-#include "rust/bindings.h"
-
-int proxy_main(const ffi::ProxyCliArgs *argsFfi);
-
-#endif
